Extracted createPatternParser
